test(smoke): add app startup smoke test
