+++ .github/workflows/flutter_ci.yml
@@ -1,0 +1,32 @@
+
+name: Flutter CI
+
+on:
+  push:
+    branches: ["main"]
+  pull_request:
+
+jobs:
+  build:
+    runs-on: ubuntu-latest
+    steps:
+      - name: Checkout
+        uses: actions/checkout@v4
+
+      - name: Set up Flutter
+        uses: subosito/flutter-action@v2
+        with:
+          flutter-version: "3.32.8"
+          channel: "stable"
+
+      - name: Flutter pub get
+        run: flutter pub get
+
+      - name: Analyze
+        run: flutter analyze --no-fatal-infos
+
+      - name: Check format (no changes allowed)
+        run: dart format --set-exit-if-changed .
+
+      - name: Run tests
+        run: flutter test
